ci: Don't target beta channel
In case there is a change in the dart sdk, we eventually cannot fix it for both the stable and beta channel. So we have to wait until the beta channel becomes stable to fix it. In the current case we have a deprecation in the newest analyzer package, which cannot be installed when not targeting the dart beta release.